[BE] 컨설턴트 평균 별점 및 리뷰 카운트 갱신 기능 추가

Closes #64

See merge request s07-webmobile1-sub2/S07P12B305!72

+++ backend/src/main/java/com/yourseason/backend/member/consultant/controller/dto/ReviewListResponse.java
@@ -15,7 +15,7 @@ public class ReviewListResponse {
     private Long reviewId;
     private String nickname;
     private String imageUrl;
-    private double star;
+    private int star;
     private String comment;
     private LocalDate createdDate;
 }
